TARGET_STM :move usb config file to target dir
