+++ targets/TARGET_STM/TARGET_STM32F4/TARGET_NUCLEO_F411RE/USBHALHost_STM_TARGET.h
