+++ targets/TARGET_STM/TARGET_STM32F7/TARGET_DISCO_F769NI/USBHAL_STM32F769NI.h
