+++ targets/TARGET_STM/TARGET_STM32L4/TARGET_DISCO_L476VG/USBHALHost_DISCOL476VG.h
