+++ targets/TARGET_STM/TARGET_STM32L4/TARGET_L476_L486/TARGET_NUCLEO_L476RG/USBHALHost_STM_TARGET.h
@@ -15,27 +15,5 @@
 * DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 */
-
-#ifdef TARGET_STM32F303ZE
-#include "USBHAL_STM32F303ZE.h"
-#endif
-
-#if defined(TARGET_STM32F207ZG) || \
-    defined(TARGET_STM32F401RE) || \
-    defined(TARGET_STM32F407VG) || \
-    defined(TARGET_STM32F411RE) || \
-    defined(TARGET_STM32F412ZG) || \
-    defined(TARGET_STM32F429ZI) || \
-    defined(TARGET_STM32F446ZE) || \
-    defined(TARGET_STM32F746ZG) || \
-    defined(TARGET_STM32F767ZI)
-#include "USBHAL_STM_144_64pins.h"
-#endif
-
-#ifdef TARGET_STM32L476VG
-#include "USBHAL_STM32L476VG.h"
-#endif
-
-#ifdef TARGET_STM32F769NI
-#include "USBHAL_STM32F769NI.h"
-#endif
+#define USBHALHOST_64pins
+#include "USBHALHost_STM_144_64pins.h"
