+++ targets/TARGET_STM/USBHALHost_STM_144_64pins.h
