+++ targets/TARGET_STM/USBHAL_STM_144_64pins.h
